// web/archive_src/types/protocol/loom.ts
export type WorkspaceMeta = {
  id: string
  name?: string | null
  created_at_ms: number
}

export type SessionInWorkspace = {
  thread_id: string
  created_at_ms: number
}

export type ThreadInWorkspace = SessionInWorkspace

export type WorkspaceListResponse = {
  type: 'workspace_list'
  id: string
  workspaces: WorkspaceMeta[]
}

export type WorkspaceCreateResponse = {
  type: 'workspace_create'
  id: string
  workspace_id: string
}

export type WorkspaceSessionListResponse = {
  type: 'workspace_thread_list'
  id: string
  workspace_id: string
  threads: SessionInWorkspace[]
}

export type WorkspaceSessionAddResponse = {
  type: 'workspace_thread_add'
  id: string
  workspace_id: string
  thread_id: string
}

export type WorkspaceSessionRemoveResponse = {
  type: 'workspace_thread_remove'
  id: string
  workspace_id: string
  thread_id: string
}